Cleanup __main__.py even further
